Use existing collection_names
