+++ crates/connectors/ndc-postgres/src/schema.rs
@@ -9,7 +9,6 @@ use super::configuration;
 use ndc_sdk::connector;
 use ndc_sdk::models;
 use query_engine_metadata::metadata;
-use query_engine_sql::sql::ast;
 use query_engine_translation::translation::mutation::{delete, generate};
 
 /// Get the connector's schema.
@@ -272,27 +271,23 @@ fn type_to_type(typ: &metadata::Type) -> models::Type {
     }
 }
 
+/// Turn our different `Mutation` items into `ProcedureInfo`s to be output in the schema
 fn mutation_to_procedure(name: &String, mutation: &generate::Mutation) -> models::ProcedureInfo {
     match mutation {
         generate::Mutation::DeleteMutation(delete) => delete_to_procedure(name, delete),
     }
 }
 
+/// given a `DeleteMutation`, turn it into a `ProcedureInfo` to be output in the schema
 fn delete_to_procedure(name: &String, delete: &delete::DeleteMutation) -> models::ProcedureInfo {
     match delete {
         delete::DeleteMutation::DeleteByKey {
             by_column,
             description,
-            table_name: ast::TableName(table_name),
-            schema_name: ast::SchemaName(schema_name),
+            collection_name,
+            ..
         } => {
             let mut arguments = BTreeMap::new();
-
-            let result_type = if schema_name == "public" {
-                table_name.to_string()
-            } else {
-                format!("{}_{}", schema_name, table_name)
-            };
 
             arguments.insert(
                 by_column.name.clone(),
@@ -306,7 +301,9 @@ fn delete_to_procedure(name: &String, delete: &delete::DeleteMutation) -> models
                 name: name.to_string(),
                 description: Some(description.to_string()),
                 arguments,
-                result_type: models::Type::Named { name: result_type },
+                result_type: models::Type::Named {
+                    name: collection_name.to_string(),
+                },
             }
         }
     }
